# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: SHIELD API
  version: 1.0.0
  description: Smart Housing Infrastructure and Entry Log Digitalization API
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /maintenance-charges/{id}:
    get:
      tags: [Billing]
      summary: Get maintenance charge by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Maintenance charge fetched}
    put:
      tags: [Billing]
      summary: Update maintenance charge
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Maintenance charge updated}
    delete:
      tags: [Billing]
      summary: Delete maintenance charge
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Maintenance charge deleted}
...
  /maintenance-charges/cycle/{cycleId}:
    get:
      tags: [Billing]
      summary: List maintenance charges by billing cycle
      parameters:
        - in: path
          name: cycleId
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Maintenance charges by cycle fetched}
  /maintenance-charges/unit/{unitId}:
    get:
      tags: [Billing]
      summary: List maintenance charges by unit
      parameters:
        - in: path
          name: unitId
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Maintenance charges by unit fetched}
...
  /late-fee-rules:
    get:
      tags: [Billing]
      summary: List late fee rules
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Late fee rules fetched}
    post:
      tags: [Billing]
      summary: Create late fee rule
      responses:
        '200': {description: Late fee rule created}
  /late-fee-rules/{id}:
    get:
      tags: [Billing]
      summary: Get late fee rule by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Late fee rule fetched}
    put:
      tags: [Billing]
      summary: Update late fee rule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Late fee rule updated}
    delete:
      tags: [Billing]
      summary: Delete late fee rule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Late fee rule deleted}
  /late-fee-rules/{id}/activate:
    post:
      tags: [Billing]
      summary: Activate late fee rule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Late fee rule activated}
  /late-fee-rules/{id}/deactivate:
    post:
      tags: [Billing]
      summary: Deactivate late fee rule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Late fee rule deactivated}
  /payments:
    get:
      tags: [Billing]
      summary: List payments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Payments fetched}
    post:
      tags: [Billing]
      summary: Create payment
      responses:
        '200': {description: Payment created}
  /payments/{id}:
    get:
      tags: [Billing]
      summary: Get payment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Payment fetched}
  /payments/invoice/{invoiceId}:
    get:
      tags: [Billing]
      summary: List payments by invoice
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Payments by invoice fetched}
  /payments/unit/{unitId}:
    get:
      tags: [Billing]
      summary: List payments by unit
      parameters:
        - in: path
          name: unitId
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Payments by unit fetched}
  /payments/{id}/receipt:
    get:
      tags: [Billing]
      summary: Get payment receipt
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Payment receipt fetched}
...
  /payments/{id}/refund:
    post:
      tags: [Billing]
      summary: Refund payment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Payment refunded}
...
  /payments/webhook/{provider}:
    post:
      tags: [Billing]
      summary: Process provider webhook callback with signature validation
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
        - in: header
          name: X-SHIELD-SIGNATURE
          required: false
          schema:
            type: string
      responses:
        '200': {description: Payment webhook processed}
  /payments/transaction/{transactionRef}:
    get:
      tags: [Billing]
      summary: Get payment gateway transaction by reference
      parameters:
        - in: path
          name: transactionRef
          required: true
          schema:
            type: string
      responses:
        '200': {description: Payment gateway transaction fetched}
...
  /asset-depreciation/asset/{assetId}:
    get:
      tags: [Asset]
      summary: List asset depreciation by asset
      parameters:
        - in: path
          name: assetId
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Asset depreciation by asset fetched}
...
  /asset-depreciation/year/{year}:
    get:
      tags: [Asset]
      summary: List asset depreciation by year
      parameters:
        - in: path
          name: year
          required: true
          schema: {type: integer}
      responses:
        '200': {description: Asset depreciation by year fetched}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Sort:
      in: query
      name: sort
      schema:
        type: string
        example: createdAt,desc
...